PersonalDataApi/requirements.txt
django >2
django-filter
django-cors-headers
djongo
graphene
graphene-django
graphql-core
django-graphql-jwt